Move `Staged` repository to `adapters` module
