+++ packages/ploys/src/repository/adapters/mod.rs
@@ -1,1 +1,2 @@
+pub mod staged;
 pub mod subdirectory;
