+++ packages/ploys/src/repository/adapters/staged.rs
@@ -1,5 +1,3 @@
-mod drain;
-
 use std::borrow::Cow;
 use std::collections::BTreeMap;
 
@@ -9,8 +7,6 @@ use relative_path::{RelativePath, RelativePathBuf};
 
 use crate::repository::path::prepare_path;
 use crate::repository::{Repository, Stage};
-
-use self::drain::Drain;
 
 /// A repository adapter for staging changes.
 #[derive(Clone)]
@@ -38,6 +34,17 @@ impl<T> Staged<T> {
 
     /// Drains the staged files.
     pub(crate) fn drain(&mut self) -> impl Iterator<Item = (RelativePathBuf, Option<Bytes>)> {
+        /// An iterator that drains a file map.
+        pub struct Drain<'a>(pub &'a mut BTreeMap<RelativePathBuf, Option<Bytes>>);
+
+        impl<'a> Iterator for Drain<'a> {
+            type Item = (RelativePathBuf, Option<Bytes>);
+
+            fn next(&mut self) -> Option<Self::Item> {
+                self.0.pop_first()
+            }
+        }
+
         Drain(&mut self.files).fuse()
     }
 }
