+++ packages/ploys/src/repository/types/mod.rs
@@ -1,4 +1,3 @@
-pub mod staged;
 pub mod staging;
 
 #[cfg(feature = "fs")]
